Extracted class to own file
